Compatibility for ESP8266

+++ ESPixelStick/src/ESPixelStick.h
@@ -31,7 +31,8 @@
 #else
 #	error "Unsupported CPU type"
 #endif
-#include "esp_err.h"
+
+#include "backported.h"
 
 #define ARDUINOJSON_USE_LONG_LONG 1
 
@@ -126,7 +127,7 @@ inline esp_err_t saferRgbToHtmlColorString(char (&output)[N], uint8_t r, uint8_t
     static_assert(N >= 8);
     int wouldHaveWrittenChars = snprintf(output, N, "#%02x%02x%02x", r, g, b);
     // TODO: assert ((wouldHaveWrittenChars > 0) && (wouldHaveWrittenChars < N))
-    if ((wouldHaveWrittenChars > 0) && (wouldHaveWrittenChars < N)) {
+    if ((wouldHaveWrittenChars > 0) && (((size_t)wouldHaveWrittenChars) < N)) {
         return ESP_OK;
     }
     return ESP_FAIL;
@@ -149,7 +150,7 @@ inline esp_err_t saferSecondsToFormattedMinutesAndSecondsString(char (&output)[N
     uint8_t  s = seconds % 60u;
     int wouldHaveWrittenChars = snprintf(output, N, "%u:%02u", m, s);
     // TODO: assert ((wouldHaveWrittenChars > 0) && (wouldHaveWrittenChars < N))
-    if ((wouldHaveWrittenChars > 0) && (wouldHaveWrittenChars < N)) {
+    if ((wouldHaveWrittenChars > 0) && (((size_t)wouldHaveWrittenChars) < N)) {
         return ESP_OK;
     }
     return ESP_FAIL;
